NSD Service Starting Function
NSD Service is necessary to use NDI on Android. NDI is an video and audio transer technology from NewTek and it is widely used especially in virtual production. NewTek released an official Unreal Engine plugin but it doesn't work on Android. So, I created an Android one. I use this plugin and nsd implementation to finish it.